bunch of formatting and minor structure changes, and socket status bug fix
This commit changes public API (see the SocketStatus enum)

The only functional changes in this commit is that this fixes a bug
where the socket status register was interpreted as a bitfield, and when
we want to change the socket's status, we now wait for the desired
state, instead of waiting until we leave the current state.

Example of bitfield bug:

    while self.0.read_u8(socket.at(SocketRegister::Status)).await?
        & SocketStatus::Closed as u8
        > 0
    { yield_now().await; }

should be:

    while self.0.read_u8(socket.at(SocketRegister::Status)).await?
        == SocketStatus::Closed as u8
    { yield_now().await; }

and the other change: (here we're opening a socket)

    while self.0.read_u8(socket.at(SocketRegister::Status)).await?
        != SocketStatus::Init as u8
    { yield_now().await; }

Furthermore, there is now a socket_status function to reduce this to:

    while self.0.socket_status(socket).await? != SocketStatus::Init {
        yield_now().await;
    }

further increasing readability. Eventually we should probably introduce
more functions like this for other registers. Also, it might be better
for these functions to live on some kind of wrapper impl/struct,
especially if we change the current situation of "async traits" to also
use wrapper structs instead. (so we don't need GATs)